refactor: rename to chevereto_crawler.py
